refactored useFirestore into two seperate hooks
