+++ src/utils/gitHubApi.js
@@ -6,14 +6,14 @@ export const fetchUser = async (userQuery) => {
    return ({
       status: res.status,
       id: user.id,
-      login: user.login || 'N/A',
-      html_url: user.html_url,
+      login: user.login ,
+      html_url: user.html_url ,
       _name: user.name || 'N/A',
-      public_repos: user.public_repos,
-      public_gists: user.public_gists,
-      followers: user.followers,
-      following: user.following,
-      created_at: user.created_at
+      public_repos: user.public_repos ,
+      public_gists: user.public_gists ,
+      followers: user.followers ,
+      following: user.following ,
+      created_at: user.created_at ,
    })
  
 }
