+++ src/utils/mungeQuery.js
@@ -1,0 +1,4 @@
+export const mungeQuery = (str) => {
+   const washed = str.replace(/[.,/#!$@+<>?"'%^&*;:{}=\-_`~()]/g,"")
+   return washed.replace(/\s+/g, '').trim()
+}
